Retirado a implementação de applicationRepository do Controller e incluído no Service; Criado Indexes no Schemas.sql; Alterado o Delete para ser apenas um SoftDelete;
