+++ src/main/java/br/com/codenation/v1/errorManager/application/ApplicationService.java
@@ -4,7 +4,15 @@ import br.com.codenation.v1.errorManager.exception.UserNaoEncontradoException;
 import br.com.codenation.v1.errorManager.user.User;
 import br.com.codenation.v1.errorManager.user.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Example;
+import org.springframework.data.domain.ExampleMatcher;
 import org.springframework.stereotype.Service;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.server.ResponseStatusException;
+
+import java.util.List;
+
+import static org.springframework.http.HttpStatus.NOT_FOUND;
 
 @Service
 public class ApplicationService {
@@ -14,6 +22,15 @@ public class ApplicationService {
 
     @Autowired
     UserRepository userRepository;
+
+    public List<Application> findApplications(Application filtro){
+        ExampleMatcher matcher = ExampleMatcher.matching()
+                .withIgnoreCase()
+                .withStringMatcher(ExampleMatcher.StringMatcher.CONTAINING);
+        Example criteria = Example.of(filtro, matcher);
+
+        return applicationRepository.findAll(criteria);
+    }
 
     public Application saveApplication(ApplicationDTO dto){
         User user = convertUser(dto.getUserId());
@@ -28,5 +45,14 @@ public class ApplicationService {
     private User convertUser(Long id){
         return userRepository.findById(id)
                 .orElseThrow(() -> new UserNaoEncontradoException());
+    }
+
+    public void deleteApplication(Long id){
+        applicationRepository.findById(id)
+                .map(a -> {
+                    a.setActive(false);
+                    applicationRepository.save(a);
+                    return a;
+                }).orElseThrow(() -> new UserNaoEncontradoException());
     }
 }
