+++ src/main/java/br/com/codenation/v1/errorManager/user/User.java
@@ -1,5 +1,6 @@
 package br.com.codenation.v1.errorManager.user;
 
+import br.com.codenation.v1.errorManager.application.Application;
 import br.com.codenation.v1.errorManager.enums.Profile;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -7,22 +8,14 @@ import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
-import javax.persistence.CollectionTable;
-import javax.persistence.Column;
-import javax.persistence.ElementCollection;
-import javax.persistence.Entity;
-import javax.persistence.EntityListeners;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
+import javax.persistence.*;
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
 
 import java.time.LocalDateTime;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Objects;
 import java.util.Set;
 import java.util.stream.Collectors;
@@ -60,7 +53,6 @@ public class User {
   @ElementCollection(fetch = FetchType.EAGER)
   @CollectionTable(name = "PROFILES")
   private Set<Integer> profiles = new HashSet<>();
-
 
   public User() {
     addProfile(Profile.USER);
