+++ src/main/resources/schemas.sql
@@ -1,7 +1,3 @@
-<!--CREATE DATABASE ERRORMANAGERAPPLICATION;-->
-
-<!--USE ERRORMANAGERAPPLICATION;-->
-
 CREATE TABLE USERS
 (
       ID INTEGER NOT NULL PRIMARY KEY AUTO_INCREMENT
@@ -9,18 +5,26 @@ CREATE TABLE USERS
     , PASSWORD VARCHAR(64)
     , CREATED_AT TIMESTAMP DEFAULT NOW()
     , UPDATED_AT TIMESTAMP
-)
+);
+
+CREATE INDEX IDX_USERS_USERNAME ON USERS (lower(USERNAME));
 
 CREATE TABLE APPLICATIONS
 (
-      ID INTEGER PRIMARY KEY AUTO_INCREMENT
+      ID INTEGER NOT NULL PRIMARY KEY AUTO_INCREMENT
     , NAME VARCHAR(255)
     , CREATED_AT TIMESTAMP DEFAULT NOW()
     , USER_ID INTEGER REFERENCES USERS (ID)
+    , IS_ACTIVE BOOLEAN DEFAULT TRUE
 );
+
+CREATE INDEX IDX_APPLICATIONS_NAME ON APPLICATIONS (lower(NAME));
 
 CREATE TABLE PROFILES
 (
     USER_ID BIGINT NOT NULL REFERENCES USERS (ID)
     PROFILES INTEGER
 );
+
+CREATE INDEX IDX_PROFILES_USER_ID ON PROFILES (USER_ID);
+
